move external module to dim-extern-api (#541)
