+++ dim/Cargo.toml
@@ -21,24 +21,21 @@ nightfall = { git = "https://github.com/Dusk-Labs/nightfall", tag = "0.3.12-rc4"
 dim-auth = { path = "../dim-auth" }
 dim-database = { path = "../dim-database" }
 dim-events = { path = "../dim-events" }
+dim-extern-api = { path = "../dim-extern-api" }
 
 serde = { version = "^1.0.125", default-features = false, features = ["derive", "std", "rc"] }
 serde_derive = "^1.0.125"
 serde_json = "^1.0.64"
 
-anitomy = "0.1.2"
-async-recursion = "0.3.2"
 async-trait = "0.1.50"
 bytes = "1.0.1"
 cfg-if = "1.0.0"
 chrono = { version = "0.4.19", features = ["serde"] }
-dashmap = "5.3.3"
 dia-i18n = "0.10.0"
 displaydoc = "0.2.3"
 dominant_color = "0.3.0"
 futures = "0.3.14"
 fuzzy-matcher = "0.3.7"
-governor = "0.5.0"
 http = "^0.2.3"
 image = "0.24.3"
 itertools = "0.10.3"
@@ -47,7 +44,6 @@ notify = "5.0.0"
 once_cell = "1.8.0"
 parking_lot = "0.12.0"
 percent-encoding = "2.1.0"
-priority-queue = "1.2.0"
 rand = { version = "0.8.5", features = ["small_rng"] }
 reqwest = { version = "0.11.0", features = ["json", "rustls-tls", "brotli"], default-features = false }
 rust-embed = "^5.9.0"
@@ -56,7 +52,6 @@ structopt = "0.3.21"
 thiserror = "1.0.30"
 tokio = { version = "1", features = ["rt", "signal", "full", "tracing"] }
 toml = "0.5.8"
-torrent-name-parser = "0.6.3"
 tracing = "0.1.32"
 tracing-appender = "0.2.0"
 tracing-subscriber = { version = "^0.3.10", features = ["fmt", "env-filter", "json"] }
@@ -70,7 +65,6 @@ xtra = { version = "0.5.1", features = ["with-tokio-1"] }
 # FIXME: Remove when we get rid of xtra_proc
 new_xtra = { package = "xtra", git = "https://github.com/Restioson/xtra", features = ["with-tokio-1"] }
 url = "2.2.2"
-retry-block = "1.0.0"
 hyper = "0.14.20"
 rusqlite = { version = "0.27.0", features = ["hooks"] }
 libsqlite3-sys = { version = "^0.24.0" }
